+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/KickedChatMember.kt
@@ -13,6 +13,5 @@ data class KickedChatMember(
 ) : BannedChatMember {
     @SerialName(statusField)
     @Required
-    override val status: ChatMember.Status
-        get() = ChatMember.Status.Kicked
+    override val status: ChatMember.Status = ChatMember.Status.Kicked
 }