+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/MemberChatMemberImpl.kt
@@ -11,6 +11,5 @@ data class MemberChatMemberImpl(
 ) : MemberChatMember {
     @SerialName(statusField)
     @Required
-    override val status: ChatMember.Status
-        get() = ChatMember.Status.Member
+    override val status: ChatMember.Status = ChatMember.Status.Member
 }