+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # TelegramBotAPI changelog
 
+## 9.2.1
+
+* `Version`:
+  * `Ktor`: `2.3.3` -> `2.3.4`
+* `Core`:
+  * All `ChatMember` inheritors have fixes `status` field
+
 ## 9.2.0
 
 **Add support of [Telegram Bots API 6.9](https://core.telegram.org/bots/api-changelog#september-22-2023)**